Added checks to updateTask and updated other updates

+++ src/main/java/pw/pap/service/TaskService.java
@@ -1,5 +1,6 @@
 package pw.pap.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 
 import java.time.LocalDateTime;
@@ -65,6 +66,40 @@ public class TaskService {
     public Task updateTask(Long taskId, Task updatedTask) {
         Task existingTask = taskRepository.findById(taskId)
                 .orElseThrow(() -> new IllegalArgumentException("Task not found"));
+
+        if(!updatedTask.getId().equals(existingTask.getId())){
+            throw new IllegalArgumentException("Cannot change task id");
+        }
+
+        if(updatedTask.getTitle().isEmpty()){
+            throw new IllegalArgumentException("Task title cannot be empty");
+        }
+
+        if(!updatedTask.getTaskCreationDate().equals(existingTask.getTaskCreationDate())){
+            throw new IllegalArgumentException("Cannot change task creation date");
+        }
+
+        if(!updatedTask.getTaskDeadline().equals(existingTask.getTaskDeadline())){
+            LocalDateTime currentDate = LocalDateTime.now();
+            if (updatedTask.getTaskDeadline().isBefore(currentDate)) {
+                throw new IllegalArgumentException("New deadline must be after current time");
+            }
+        }
+
+        if(!updatedTask.getAssignees().equals(existingTask.getAssignees())){
+            for (User assignee : updatedTask.getAssignees()) {
+                userRepository.findById(assignee.getId())
+                        .orElseThrow(() -> new EntityNotFoundException("One of new assignees not in database"));
+            }
+        }
+
+        if(!updatedTask.getCreator().getId().equals(existingTask.getCreator().getId())){
+            throw new IllegalArgumentException("Cannot change task creator");
+        }
+
+        if(!updatedTask.getProject().getId().equals(existingTask.getProject().getId())){
+            throw new IllegalArgumentException("Cannot change task project");
+        }
 
         taskRepository.deleteById(taskId);
         updatedTask.setId(taskId);
